cleanup placement of execution details
